Revamp file structure to match advised type
