+++ mobile/.gitattributes
