+++ mobile/Guardian_Angel/__tests__/App.js
