+++ mobile/Guardian_Angel/android/app/src/main/AndroidManifest.xml
