+++ mobile/Guardian_Angel/android/app/src/main/java/com/guardian_angel/MainActivity.java
