+++ mobile/Guardian_Angel/android/app/src/main/res/values/styles.xml
