+++ mobile/Guardian_Angel/android/gradle.properties
