+++ mobile/Guardian_Angel/android/settings.gradle
