+++ mobile/Guardian_Angel/ios/Guardian_Angel/AppDelegate.h
