+++ mobile/Guardian_Angel/ios/Guardian_Angel/Images.xcassets/AppIcon.appiconset/Contents.json
