+++ mobile/app.json
