+++ mobile/src/screens/messages/index_item.js
@@ -1,0 +1,22 @@
+import React from 'react';
+import {
+  Platform,
+  StyleSheet,
+  Text,
+  View
+} from 'react-native';
+
+export default class MessageIndexItem extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = {};
+  }
+
+  render() {
+    return (
+      <View>
+        <Text>This is a text message</Text>
+      </View>
+    );
+  }
+}
